backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.0
python-multipart==0.0.9
pydantic==2.7.1
SQLAlchemy==2.0.30
pytesseract==0.3.10
pdf2image==1.17.0
google-generativeai==0.7.2
Pillow==10.3.0
